# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:3030/api/v2","description":"Generated server url"}],"paths":{"/user/add/{username}":{"post":{"tags":["user-controller"],"operationId":"addAnime","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnimeDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/auth/register":{"post":{"tags":["auth-controller"],"operationId":"newUserRegistration","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewUserDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/auth/login":{"post":{"tags":["auth-controller"],"operationId":"authenticateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserAuthDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/auth/checkuser/{username}":{"post":{"tags":["auth-controller"],"operationId":"checkUsername","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/user/update/{username}":{"patch":{"tags":["user-controller"],"operationId":"editAnime","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnimeDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/user/view/{username}":{"get":{"tags":["user-controller"],"operationId":"getAllContent","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AnimeDetailsDto"}}}}}}}},"/user/details/{username}":{"get":{"tags":["user-controller"],"operationId":"getUser","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/content/show/{contentId}":{"get":{"tags":["content-controller"],"operationId":"getContent","parameters":[{"name":"contentId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ContentDetailsDto"}}}}}}},"/content/search":{"get":{"tags":["content-controller"],"operationId":"getContentOnQuery","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ContentDetailsDto"}}}}}}}},"/content/recommendation/{contentId}":{"get":{"tags":["content-controller"],"operationId":"getRecommendations","parameters":[{"name":"contentId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ContentDetailsDto"}}}}}}}},"/content/list":{"get":{"tags":["content-controller"],"operationId":"getContentList","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ContentDetailsDto"}}}}}}}},"/user/delete/{username}":{"delete":{"tags":["user-controller"],"operationId":"deleteAnime","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContentIdDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}}},"components":{"schemas":{"AnimeDto":{"type":"object","properties":{"contentId":{"type":"string"},"status":{"type":"string"},"score":{"type":"number","format":"double"},"episodeProgress":{"type":"integer","format":"int32"},"review":{"type":"string"}}},"ApiResponse":{"type":"object","properties":{"localDate":{"type":"string","format":"date"},"message":{"type":"string"}}},"NewUserDto":{"type":"object","properties":{"email":{"type":"string"},"userName":{"type":"string"},"password":{"type":"string"}}},"UserResponse":{"type":"object","properties":{"userName":{"type":"string"},"email":{"type":"string"},"token":{"type":"string"}}},"UserAuthDto":{"type":"object","properties":{"userName":{"type":"string"},"password":{"type":"string"}}},"AnimeDetailsDto":{"type":"object","properties":{"contentId":{"type":"string"},"title":{"type":"string"},"type":{"type":"string","enum":["TV","MOVIE","OVA","ONA","SPECIAL","UNKNOWN"]},"picture":{"type":"string"},"thumbnail":{"type":"string"},"season":{"$ref":"#/components/schemas/Season"},"status":{"type":"string","enum":["COMPLETED","DROPPED","PAUSED","PLAN_TO_WATCH","REWATCHING","WATCHING"]},"contentStatus":{"type":"string","enum":["FINISHED","ONGOING","UPCOMING","UNKNOWN"]},"episodeProgress":{"type":"integer","format":"int32"},"totalEpisodes":{"type":"integer","format":"int32"},"score":{"type":"number","format":"double"},"review":{"type":"string"}}},"Season":{"type":"object","properties":{"season":{"type":"string","enum":["SPRING","SUMMER","FALL","WINTER","UNDEFINED"]},"year":{"type":"integer","format":"int32"}}},"ContentDetailsDto":{"type":"object","properties":{"contentId":{"type":"string"},"title":{"type":"string"},"picture":{"type":"string"},"thumbnail":{"type":"string"},"description":{"type":"string"},"totalEpisodes":{"type":"integer","format":"int32"},"type":{"type":"string","enum":["TV","MOVIE","OVA","ONA","SPECIAL","UNKNOWN"]},"contentStatus":{"type":"string","enum":["FINISHED","ONGOING","UPCOMING","UNKNOWN"]},"score":{"type":"number","format":"double"}}},"ContentIdDto":{"type":"object","properties":{"contentId":{"type":"string"}}}}}}